requirements.txt
aiogram==3.7.0
aiohttp==3.9.5
beautifulsoup4==4.12.3
lxml==5.2.2
apscheduler==3.10.4
telethon==1.36.0
psycopg2-binary==2.9.9
